requirements.txt
aiogram==3.22.0
aiosqlite==0.21.0
edge-tts==7.2.1
SpeechRecognition==3.14.4
python-dotenv==1.2.1
chatgpt-md-converter==0.3.10
tzdata==2025.2
loguru==0.7.3
pillow==12.0.0
pytest==9.0.1
pytest-asyncio>=1.3.0
colorama==0.4.6
llama-cpp-py>=0.1.18
# git+https://github.com/sergey21000/llama-cpp-py.git